Follow-up for r20511: really use the proper format specification.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-5-patches@20512 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/agentx/protocol.c
@@ -1172,8 +1172,8 @@ agentx_parse_oid(u_char * data, size_t * length, int *inc,
      */
     tmp_oid_len = (prefix ? n_subid + 5 : n_subid);
     if (*oid_len < tmp_oid_len) {
-        DEBUGMSGTL(("agentx", "Oversized Object ID (buf=%ld pdu=%d)\n", *oid_len,
-                 tmp_oid_len));
+        DEBUGMSGTL(("agentx", "Oversized Object ID (buf=%d pdu=%d)\n",
+		    (int) *oid_len, tmp_oid_len));
         DEBUGINDENTLESS();
         return NULL;
     }
